Move Writers to its own file.
